Add init version of new stripe plugin. deprecate the old one
