+++ saleor/payment/gateways/stripe/consts.py
@@ -1,12 +1,9 @@
-
-
-PLUGIN_ID = "mirumee.payments.stripe"
+PLUGIN_ID = "saleor.payments.stripe"
 PLUGIN_NAME = "Stripe"
 WEBHOOK_PATH = "webhooks/"
 WEBHOOK_EVENTS = [
     "payment_intent.payment_failed",
     "payment_intent.succeeded",
-    "payment_intent.amount_capturable_updated",
     "payment_intent.processing",
 ]
 METADATA_IDENTIFIER = "saleor-domain"
@@ -16,10 +13,7 @@ ACTION_REQUIRED_STATUSES = [
     "requires_confirmation",
     "requires_action",
 ]
-FAILED_STATUSES = [
-    "requires_payment_method"
-    "canceled"
-]
+FAILED_STATUSES = ["requires_payment_method" "canceled"]
 
 SUCCESS_STATUSES = "succeeded"
