+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_authorize.yaml
