+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_authorize_3d_secure.yaml
