+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_authorize_and_save_customer_id.yaml
