+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_authorize_error_response.yaml
