+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_authorize_with_customer_id.yaml
