+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_authorize_without_capture.yaml
