+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_capture.yaml
