+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_capture_3d_secure.yaml
