+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_capture_error_response.yaml
