+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_confirm__intent.yaml
