+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_confirm_error_response.yaml
