+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_list_customer_sources.yaml
