+++ saleor/payment/gateways/stripe/deprecated/tests/cassettes/test_void_error_response.yaml
