+++ saleor/payment/gateways/stripe/deprecated/tests/test_stripe.py
@@ -4,9 +4,9 @@ from math import isclose
 
 import pytest
 
-from .... import ChargeStatus
-from ....interface import CustomerSource, GatewayConfig, PaymentMethodInfo
-from ....utils import create_payment_information
+from ..... import ChargeStatus
+from .....interface import CustomerSource, GatewayConfig, PaymentMethodInfo
+from .....utils import create_payment_information
 from .. import (
     TransactionKind,
     _get_client,
