+++ saleor/payment/gateways/stripe/plugin.py
@@ -1,60 +1,81 @@
-from typing import TYPE_CHECKING, List
+import logging
+from typing import TYPE_CHECKING, Tuple
 
-from saleor.plugins.base_plugin import BasePlugin, ConfigurationTypeField
+from django.contrib.sites.models import Site
+from django.core.exceptions import ValidationError
+from django.core.handlers.wsgi import WSGIRequest
+from django.http import HttpResponse, HttpResponseNotFound
+from django.http.request import split_domain_port
+from stripe.stripe_object import StripeObject
 
+from ....graphql.core.enums import PluginErrorCode
+from ....plugins.base_plugin import BasePlugin, ConfigurationTypeField
+from ... import TransactionKind
+from ...interface import GatewayConfig, GatewayResponse, PaymentData
+from ...models import Transaction
+from ...utils import price_from_minor_unit
 from ..utils import get_supported_currencies, require_active_plugin
-from . import (
-    GatewayConfig,
-    authorize,
-    capture,
-    list_client_sources,
-    process_payment,
-    refund,
-    void,
+from .stripe_api import (
+    create_payment_intent,
+    is_secret_api_key_valid,
+    retrieve_payment_intent,
+    retrieve_webhook,
+    subscribe_to_webhook,
 )
-
-GATEWAY_NAME = "Stripe"
+from .webhooks import handle_webhook
 
 if TYPE_CHECKING:
     # flake8: noqa
-    from ...interface import CustomerSource
-    from . import GatewayResponse, PaymentData
+    from ....plugins.models import PluginConfiguration
+
+from .consts import (
+    ACTION_REQUIRED_STATUSES,
+    PLUGIN_ID,
+    PLUGIN_NAME,
+    PROCESSING_STATUS,
+    SUCCESS_STATUSES,
+    WEBHOOK_PATH,
+)
+
+logger = logging.getLogger(__name__)
 
 
 class StripeGatewayPlugin(BasePlugin):
-    PLUGIN_NAME = GATEWAY_NAME
-    PLUGIN_ID = "mirumee.payments.stripe"
+    PLUGIN_NAME = PLUGIN_NAME
+    PLUGIN_ID = PLUGIN_ID
     DEFAULT_CONFIGURATION = [
-        {"name": "Public API key", "value": None},
-        {"name": "Secret API key", "value": None},
-        {"name": "Store customers card", "value": False},
-        {"name": "Automatic payment capture", "value": True},
-        {"name": "Supported currencies", "value": ""},
+        {"name": "public_api_key", "value": None},
+        {"name": "secret_api_key", "value": None},
+        {"name": "store_customers_cards", "value": False},
+        {"name": "automatic_payment_capture", "value": True},
+        {"name": "supported_currencies", "value": ""},
+        {"name": "webhook_endpoint_id", "value": None},
+        {"name": "webhook_secret_key", "value": None},
     ]
 
     CONFIG_STRUCTURE = {
-        "Public API key": {
-            "type": ConfigurationTypeField.SECRET,
+        "public_api_key": {
+            "type": ConfigurationTypeField.STRING,
             "help_text": "Provide Stripe public API key.",
             "label": "Public API key",
         },
-        "Secret API key": {
+        "secret_api_key": {
             "type": ConfigurationTypeField.SECRET,
             "help_text": "Provide Stripe secret API key.",
             "label": "Secret API key",
         },
-        "Store customers card": {
+        "store_customers_cards": {
             "type": ConfigurationTypeField.BOOLEAN,
             "help_text": "Determines if Saleor should store cards on payments "
             "in Stripe customer.",
             "label": "Store customers card",
         },
-        "Automatic payment capture": {
+        "automatic_payment_capture": {
             "type": ConfigurationTypeField.BOOLEAN,
             "help_text": "Determines if Saleor should automaticaly capture payments.",
             "label": "Automatic payment capture",
         },
-        "Supported currencies": {
+        "supported_currencies": {
             "type": ConfigurationTypeField.STRING,
             "help_text": "Determines currencies supported by gateway."
             " Please enter currency codes separated by a comma.",
@@ -63,69 +84,269 @@ class StripeGatewayPlugin(BasePlugin):
     }
 
     def __init__(self, *args, **kwargs):
+
+        # Webhook details are not listed in CONFIG_STRUCTURE as user input is not
+        # required here
+        plugin_configuration = kwargs.get("configuration")
+        raw_configuration = {
+            item["name"]: item["value"] for item in plugin_configuration
+        }
+        webhook_endpoint_id = raw_configuration.get("webhook_endpoint_id")
+        webhook_secret = raw_configuration.get("webhook_secret_key")
+
         super().__init__(*args, **kwargs)
         configuration = {item["name"]: item["value"] for item in self.configuration}
         self.config = GatewayConfig(
-            gateway_name=GATEWAY_NAME,
-            auto_capture=configuration["Automatic payment capture"],
-            supported_currencies=configuration["Supported currencies"],
+            gateway_name=PLUGIN_NAME,
+            auto_capture=configuration["automatic_payment_capture"],
+            supported_currencies=configuration["supported_currencies"],
             connection_params={
-                "public_key": configuration["Public API key"],
-                "private_key": configuration["Secret API key"],
+                "public_api_key": configuration["public_api_key"],
+                "secret_api_key": configuration["secret_api_key"],
+                "webhook_id": webhook_endpoint_id,
+                "webhook_secret": webhook_secret,
             },
-            store_customer=configuration["Store customers card"],
+            # FIXME define if we are able to implement support for this
+            store_customer=configuration["store_customers_cards"],
         )
 
-    def _get_gateway_config(self):
-        return self.config
+    def webhook(self, request: WSGIRequest, path: str, previous_value) -> HttpResponse:
+        config = self.config
+        if path.startswith(WEBHOOK_PATH, 1):  # 1 as we don't check the '/'
+            return handle_webhook(request, config)
+        return HttpResponseNotFound()
 
     @require_active_plugin
-    def authorize_payment(
-        self, payment_information: "PaymentData", previous_value
-    ) -> "GatewayResponse":
-        return authorize(payment_information, self._get_gateway_config())
-
-    @require_active_plugin
-    def capture_payment(
-        self, payment_information: "PaymentData", previous_value
-    ) -> "GatewayResponse":
-        return capture(payment_information, self._get_gateway_config())
-
-    @require_active_plugin
-    def refund_payment(
-        self, payment_information: "PaymentData", previous_value
-    ) -> "GatewayResponse":
-        return refund(payment_information, self._get_gateway_config())
-
-    @require_active_plugin
-    def void_payment(
-        self, payment_information: "PaymentData", previous_value
-    ) -> "GatewayResponse":
-        return void(payment_information, self._get_gateway_config())
-
-    @require_active_plugin
-    def process_payment(
-        self, payment_information: "PaymentData", previous_value
-    ) -> "GatewayResponse":
-        return process_payment(payment_information, self._get_gateway_config())
-
-    @require_active_plugin
-    def list_payment_sources(
-        self, customer_id: str, previous_value
-    ) -> List["CustomerSource"]:
-        sources = list_client_sources(self._get_gateway_config(), customer_id)
-        previous_value.extend(sources)
-        return previous_value
+    def token_is_required_as_payment_input(self, previous_value):
+        return False
 
     @require_active_plugin
     def get_supported_currencies(self, previous_value):
-        config = self._get_gateway_config()
-        return get_supported_currencies(config, GATEWAY_NAME)
+        return get_supported_currencies(self.config, PLUGIN_NAME)
 
-    @require_active_plugin
-    def get_payment_config(self, previous_value):
-        config = self._get_gateway_config()
-        return [
-            {"field": "api_key", "value": config.connection_params["public_key"]},
-            {"field": "store_customer_card", "value": config.store_customer},
-        ]
+    def process_payment(
+        self, payment_information: "PaymentData", previous_value
+    ) -> "GatewayResponse":
+
+        intent, error = create_payment_intent(
+            api_key=self.config.connection_params["secret_api_key"],
+            amount=payment_information.amount,
+            currency=payment_information.currency,
+        )
+        raw_response = None
+        client_secret = None
+        if intent:
+            client_secret = intent.client_secret
+            raw_response = intent.last_response.data
+        return GatewayResponse(
+            is_success=True if not error else False,
+            action_required=True,
+            kind=TransactionKind.ACTION_TO_CONFIRM,
+            amount=payment_information.amount,
+            currency=payment_information.currency,
+            transaction_id=intent.id if intent else "",
+            error=error,
+            raw_response=raw_response,
+            action_required_data={"client_secret": client_secret},
+        )
+
+    def confirm_payment(
+        self, payment_information: "PaymentData", previous_value
+    ) -> "GatewayResponse":
+        payment_intent_id = payment_information.token
+        api_key = self.config.connection_params["secret_api_key"]
+
+        # before we will call stripe API, let's check if the transaction object hasn't
+        # been created by webhook handler
+        payment_transaction = Transaction.objects.filter(
+            payment_id=payment_information.payment_id,
+            is_success=True,
+            action_required=False,
+            kind__in=[TransactionKind.AUTH, TransactionKind.CAPTURE],
+        ).first()
+
+        if payment_transaction:
+            return GatewayResponse(
+                is_success=True,
+                action_required=False,
+                kind=payment_transaction.kind,
+                amount=payment_transaction.amount,
+                currency=payment_transaction.currency,
+                transaction_id=payment_transaction.token,
+                error=None,
+                raw_response=payment_transaction.gateway_response,
+                transaction_already_processed=True,
+            )
+
+        payment_intent = None
+        if payment_intent_id:
+            payment_intent = retrieve_payment_intent(api_key, payment_intent_id)
+
+        kind = (
+            TransactionKind.CAPTURE
+            if self.config.auto_capture
+            else TransactionKind.AUTH
+        )
+        action_required = False
+
+        if payment_intent:
+            amount = price_from_minor_unit(
+                payment_intent.amount, payment_intent.currency
+            )
+            currency = payment_intent.currency
+
+            # payment still requires an action
+            if payment_intent.status in ACTION_REQUIRED_STATUSES:
+                kind = TransactionKind.ACTION_TO_CONFIRM
+                action_required = True
+
+            elif payment_intent.status == PROCESSING_STATUS:
+                kind = TransactionKind.PENDING
+            elif payment_intent.status == SUCCESS_STATUSES:
+                kind = (
+                    TransactionKind.CAPTURE
+                    if payment_intent.capture_method == "automatic"
+                    else TransactionKind.AUTH
+                )
+        else:
+            amount = payment_information.amount
+            currency = payment_information.currency
+
+        return GatewayResponse(
+            is_success=True if payment_intent else False,
+            action_required=action_required,
+            kind=kind,
+            amount=amount,
+            currency=currency,
+            transaction_id=payment_intent.id if payment_intent else "",
+            error=None,
+            raw_response=payment_intent.last_response.data if payment_intent else None,
+        )
+
+    @classmethod
+    def pre_save_for_existing_webhook_configuration(
+        cls, api_key, plugin_configuration: "PluginConfiguration", configuration: dict
+    ) -> Tuple[StripeObject, bool]:
+        # make sure that webhook exists on Stripe side
+        webhook = retrieve_webhook(api_key, configuration["webhook_endpoint_id"])
+
+        new_subscription = False
+
+        # remove the old webhook detail values if we were not able to fetch a
+        # webhook with the id
+        if not webhook:
+            plugin_configuration.configuration.remove(
+                {
+                    "name": "webhook_endpoint_id",
+                    "value": configuration["webhook_endpoint_id"],
+                }
+            )
+            plugin_configuration.configuration.remove(
+                {
+                    "name": "webhook_secret_key",
+                    "value": configuration["webhook_secret_key"],
+                }
+            )
+            webhook = subscribe_to_webhook(api_key)
+            new_subscription = True
+        return webhook, new_subscription
+
+    @classmethod
+    def pre_save_for_non_existing_webhook_configuration(
+        cls, api_key, plugin_configuration: "PluginConfiguration", configuration: dict
+    ) -> StripeObject:
+        # Create new webhook subscription in case when we don't have all webhook
+        # details
+        webhook = subscribe_to_webhook(api_key)
+        if configuration.get("webhook_endpoint_id"):
+            plugin_configuration.configuration.remove(
+                {
+                    "name": "webhook_endpoint_id",
+                    "value": configuration["webhook_endpoint_id"],
+                }
+            )
+        if configuration.get("webhook_secret_key"):
+            plugin_configuration.configuration.remove(
+                {
+                    "name": "webhook_secret_key",
+                    "value": configuration["webhook_secret_key"],
+                }
+            )
+        return webhook
+
+    @classmethod
+    def pre_save_plugin_configuration(cls, plugin_configuration: "PluginConfiguration"):
+        configuration = plugin_configuration.configuration
+        configuration = {item["name"]: item["value"] for item in configuration}
+        if not plugin_configuration.active:
+            return
+
+        # check saved domain. Make sure that it is not localhost domain. We are not able
+        # to subscribe to stripe webhooks with localhost.
+        domain = Site.objects.get_current().domain
+        localhost_domains = ["localhost", "127.0.0.1"]
+        domain, _ = split_domain_port(domain)
+        if not domain:
+            logger.warning(
+                "Site doesn't have defined domain. Unable to subscribe Stripe webhooks"
+            )
+            return
+        if domain in localhost_domains:
+            logger.warning(
+                "Unable to subscribe localhost domain - %s to Stripe webhooks. Stripe "
+                "webhooks require domain which will be accessible from the network.",
+                domain,
+            )
+            return
+
+        api_key = configuration["secret_api_key"]
+        webhook_id = configuration.get("webhook_endpoint_id")
+        webhook_secret = configuration.get("webhook_secret_key")
+        if webhook_id and webhook_secret:
+            webhook, new_subscription = cls.pre_save_for_existing_webhook_configuration(
+                api_key, plugin_configuration, configuration
+            )
+        else:
+            webhook = cls.pre_save_for_non_existing_webhook_configuration(
+                api_key, plugin_configuration, configuration
+            )
+            new_subscription = True
+
+        if new_subscription:
+            plugin_configuration.configuration.extend(
+                [
+                    {"name": "webhook_endpoint_id", "value": webhook.id},
+                    {"name": "webhook_secret_key", "value": webhook.secret},
+                ]
+            )
+
+    @classmethod
+    def validate_plugin_configuration(cls, plugin_configuration: "PluginConfiguration"):
+        configuration = plugin_configuration.configuration
+        configuration = {item["name"]: item["value"] for item in configuration}
+        required_fields = ["secret_api_key", "public_api_key"]
+        all_required_fields_provided = all(
+            [configuration.get(field) for field in required_fields]
+        )
+        if plugin_configuration.active:
+            if not all_required_fields_provided:
+                raise ValidationError(
+                    {
+                        field: ValidationError(
+                            "The parameter is required.",
+                            code=PluginErrorCode.REQUIRED.value,
+                        )
+                    }
+                    for field in required_fields
+                )
+
+            api_key = configuration["secret_api_key"]
+            if not is_secret_api_key_valid(api_key):
+                raise ValidationError(
+                    {
+                        "secret_api_key": ValidationError(
+                            "Secret API key is incorrect.",
+                            code=PluginErrorCode.INVALID.value,
+                        )
+                    }
+                )
